add JWT authentication state

+++ controllers/ExpenseController.php
@@ -4,18 +4,28 @@ namespace Controllers;
 
 use App\Http\Request;
 use App\Http\Response;
+use Middlewares\AuthMiddleware;
 use Model\Expenses;
 
 class ExpenseController
 {
     public function __construct(
         public Expenses $expensesModel,
+        public AuthMiddleware $authMW,
         public Request $request
     ) {}
 
     public function addExpense(): Response
     {
         $params = $this->request->post();
+
+        $userId = $this->getUserId();
+        if (empty($userId)) {
+            return (new Response)
+            ->setStatusCode(401);
+        } else {
+            $params["user_id"] = $userId;
+        }
 
         if ($this->expensesModel->create($params)) {
             return (new Response)
@@ -28,6 +38,12 @@ class ExpenseController
 
     public function deleteExpense(): Response
     {
+        $userId = $this->getUserId();
+        if (empty($userId)) {
+            return (new Response)
+            ->setStatusCode(401);
+        }
+
         $id = $this->request->post()['id'];
         if ($this->expensesModel->deleteById($id)) {
             return (new Response)
@@ -41,6 +57,12 @@ class ExpenseController
     public function updateExpense(): Response
     {
         $params = $this->request->post();
+
+        $userId = $this->getUserId();
+        if (empty($userId)) {
+            return (new Response)
+            ->setStatusCode(401);
+        }
 
         $id = $params['id'];
         unset($params['id']);
@@ -58,6 +80,14 @@ class ExpenseController
     {
         $params = $this->request->post();
 
+        $userId = $this->getUserId();
+        if (empty($userId)) {
+            return (new Response)
+            ->setStatusCode(401);
+        } else {
+            $params["user_id"] = $userId;
+        }
+
         $expenses = $this->expensesModel->read($params);
         if (!empty($expenses)) {
             return (new Response)
@@ -67,5 +97,10 @@ class ExpenseController
             return (new Response)
             ->setStatusCode(400);
         }
+    }
+
+    private function getUserId(): ?string
+    {
+        return $this->authMW->handle();
     }
 }
